Move request queues into session ownership
Constraint: RequestQueueManager is constructed and owned by SessionManager for per-session sequencing.\nRejected: Keeping request-queue under src/utils | it is session lifecycle coordination, not a domain-neutral utility.\nRejected: Adding a compatibility shim | all in-repo callers were updated and no package export exposes this path.\nConfidence: high\nScope-risk: narrow\nDirective: Keep session sequencing primitives under src/session with SessionManager.\nTested: npx jest --config jest.ci.config.js --runInBand --silent --runTestsByPath tests/session/request-queue-timers.test.ts tests/session/manager-contracts.test.ts tests/session/manager-tenant.test.ts tests/session/manager-target-creation.test.ts tests/session/manager-evict-target.test.ts tests/session/manager-ttl.test.ts tests/recovery/self-healing-wiring.test.ts tests/integration/hybrid-lightpanda.test.ts tests/integration/tenant-isolation.test.ts\nTested: npx jest --config jest.config.js --runInBand --silent --runTestsByPath tests/session/manager-ttl.test.ts tests/integration/hybrid-lightpanda.test.ts\nTested: npm run build\nTested: npm run lint\nTested: npm run lint:tier\nTested: npm run lint:repo-structure\nTested: git diff --check\nNot-tested: Full Jest suite.
