+++ docs/releases/v1.10.4.md
@@ -248,7 +248,7 @@ The integrated release branch was validated with fresh repository checks.
 ### Targeted regression coverage merged in this release
 - `tests/transports/http-batch-limits.test.ts`
 - `tests/transports/http-bearer-auth.test.ts` (expanded for fail-closed + CORS)
-- `tests/utils/request-queue-timers.test.ts`
+- `tests/session/request-queue-timers.test.ts`
 - `tests/security/dashboard-rest-authz.test.ts`
 - `tests/tools/file-upload-validation.test.ts`
 - `tests/tools/page-screenshot-guards.test.ts`
