+++ tests/integration/hybrid-lightpanda.test.ts
@@ -124,7 +124,7 @@ jest.mock('../../src/cdp/connection-pool', () => ({
   }),
 }));
 
-jest.mock('../../src/utils/request-queue', () => ({
+jest.mock('../../src/session/request-queue', () => ({
   RequestQueueManager: jest.fn().mockImplementation(() => ({
     enqueue: jest.fn((_: unknown, fn: () => unknown) => fn()),
     deleteQueue: jest.fn(),
