+++ tests/integration/tenant-isolation.test.ts
@@ -47,7 +47,7 @@ jest.mock('../../src/cdp/connection-pool', () => ({
   getCDPConnectionPool: jest.fn().mockReturnValue({}),
 }));
 
-jest.mock('../../src/utils/request-queue', () => ({
+jest.mock('../../src/session/request-queue', () => ({
   RequestQueueManager: jest.fn().mockImplementation(() => ({
     enqueue: jest.fn((_, fn) => fn()),
     deleteQueue: jest.fn(),
